+++ src/index.ts
@@ -3,8 +3,10 @@ export * from './types'
 import { resetContext } from './context'
 
 export { kea, connect } from './kea'
+
 export { useValues, useAllValues, useActions, useMountedLogic, useKea } from './react/hooks'
 export { BindLogic } from './react/bind'
+export { Provider } from './react/provider'
 
 export { resetContext, openContext, closeContext, getContext, getPluginContext, setPluginContext } from './context'
 export { getStore } from './store/store'